% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiLinx-package.R
\docType{package}
\name{EpiLinx-package}
\alias{EpiLinx-package}
\alias{EpiLinx}
\title{Interactive Visualization Tool for Nosocomial Outbreaks}
\description{
What the package does (one paragraph).
}
\details{
Some description of the package
}
\examples{
## launch_app()

## To cite this package in publications use:
citation("EpiLinx")

}
\references{
TODO: citation for article
}
\author{
\strong{Maintainer}: Anna Emilie Henius \email{ANEH@ssi.dk}

Authors:
\itemize{
  \item Matt Denwood \email{md@sund.ku.dk}
}

}
\keyword{internal}
