% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_long_2021.R
\docType{data}
\name{test_long_2021}
\alias{test_long_2021}
\title{Example dataset}
\format{
A data frame with:
\describe{
\item{CPR_male}{TODO: description}
\item{C_STATUS}{TODO: description}
\item{CarbapenemaseNGS}{TODO: description}
\item{D_STATUS_HEN_START}{TODO: description}
\item{GeografiskLokalisation}{TODO: description}
\item{IsolatnrPat}{TODO: description}
\item{KMA}{TODO: description}
\item{Lokalnr}{TODO: description}
\item{MLST1}{TODO: description}
\item{MuligPatientUdbrudsnr}{TODO: description}
\item{MuligeUdbrud}{TODO: description}
\item{Proveart}{TODO: description}
\item{Provedato}{TODO: description}
\item{Region}{TODO: description}
\item{Regionkode}{TODO: description}
\item{Rejse}{TODO: description}
\item{SSI_id}{TODO: description}
\item{SorIdentifier}{TODO: description}
\item{Species}{TODO: description}
\item{Ward}{TODO: description}
\item{alder}{TODO: description}
\item{cgMLST}{TODO: description}
\item{course_id}{TODO: description}
\item{dept_ind_dato}{TODO: description}
\item{dept_ud_dato}{TODO: description}
\item{hospitalCPO}{TODO: description}
\item{hospital_name}{TODO: description}
\item{leave}{TODO: description}
\item{name}{TODO: description}
\item{responsible_name}{TODO: description}
\item{source}{TODO: description}
}
}
\usage{
test_long_2021
}
\description{
Some fake data to play around with
}
\keyword{datasets}
