% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_long_2021.R
\docType{data}
\name{test_long_2021}
\alias{test_long_2021}
\title{Example dataset}
\format{
A data frame with:
\describe{
\item{SSI_id}{Anonymised ID}
...
}
}
\usage{
test_long_2021
}
\description{
Some fake data to play around with
}
\keyword{datasets}
